refactor: separate out Dash-specific RPCs that rely on wallet logic
In an upcoming commit, wallet variables will be deglobalized. This means
that RPCs that use wallet logic need to get ahold of WalletContext, which
only happens if they're registered as a wallet RPC (i.e. registered
through WalletLoader).

The downside of being registered as a wallet RPC is that you lose access
to NodeContext. For now, we will work around this by giving WalletContext
access to NodeContext and modify EnsureAnyNodeContext to pull it from
WalletContext.